+++ src/api/invidious.rs
@@ -138,6 +138,7 @@ impl Api for Instance {
         channel_id: &str,
         tab: ChannelTab,
         present_videos: HashSet<String>,
+        get_all: bool,
     ) -> Result<ChannelFeed> {
         let mut feed = ChannelFeed::new(channel_id);
         let videos = self.get_more_videos_helper(channel_id, tab).await?;
@@ -154,15 +155,15 @@ impl Api for Instance {
                 .all(|video| present_videos.contains(&video.video_id))
         };
 
-        let new = new_video_present(feed.get_videos(tab));
+        let mut new = new_video_present(feed.get_videos(tab));
 
         while self.continuation.is_some()
             && let Ok(videos) = self.get_more_videos_helper(channel_id, tab).await
         {
-            let new = new_video_present(&videos);
+            new = new || new_video_present(&videos);
             feed.extend_videos(videos, tab);
 
-            if new {
+            if !get_all && new {
                 return Ok(feed);
             }
         }