+++ src/client.rs
@@ -105,10 +105,10 @@ impl Client {
                     let instance = self.instance();
                     tokio::spawn(async move { refresh_channels(instance, ids).await });
                 }
-                IoEvent::LoadMoreVideos(id, tab, present_videos) => {
+                IoEvent::LoadMoreVideos(id, tab, present_videos, load_all) => {
                     let instance = self.instance();
                     tokio::spawn(async move {
-                        get_more_videos(instance, &id, tab, present_videos).await
+                        get_more_videos(instance, &id, tab, present_videos, load_all).await
                     });
                 }
                 IoEvent::FetchFormats(title, video_id, play_selected) => {
@@ -339,8 +339,9 @@ async fn get_more_videos(
     id: &str,
     tab: ChannelTab,
     present: HashSet<String>,
+    get_all: bool,
 ) -> Result<()> {
-    match instance.get_more_videos(id, tab, present).await {
+    match instance.get_more_videos(id, tab, present, get_all).await {
         Ok(feed) => {
             if feed.get_videos(tab).is_empty() {
                 emit_msg!(warning, "There are no videos to load");