+++ src/config/keys.rs
@@ -114,6 +114,7 @@ impl Default for KeyBindings {
         insert_binding!(general, "R", Command::RefreshChannels);
         insert_binding!(general, "F", Command::RefreshFailedChannels);
         insert_binding!(general, "J", Command::LoadMoreVideos);
+        insert_binding!(general, "ctrl-j", Command::LoadAllVideos);
         insert_binding!(general, "o", Command::OpenInYoutube);
         insert_binding!(general, "O", Command::OpenInInvidious);
         insert_binding!(general, "p", Command::PlayFromFormats);